midterm/bioweb/requirements.txt
Django==4.1.2
djangorestframework==3.12.4
